Introducing: Streaming
Add new modules: rtp, rtcp. The parent crate contains algebraic data
types which are generally useful. Each submodule contains types and
functions which are useful for implementing that part of the protocol.
RTCP has made more progress, with a State tracking object and code to
join a session. No tests, but function-level documentation is provided.