Adds support for member table
Part of the RTCP protocol is tracking the number and state of
participants. This commit adds initial support for keeping the member
table populated and up to date. This is done by introducing some new
types stored in a HashMap keyed on the members SSRC. Doing it this way
requires the application to call State.pkt_recv_notify() whenever it
receives a new packet in order to keep the member counts in sync.